+++ src/main/java/mc/manifestcompany/company/CompanyAction.java
@@ -1,4 +1,4 @@
-package mc.manifestcompany;
+package mc.manifestcompany.company;
 
 /**
  * Interface for interacting with Companies.
